+++ src/main/java/de/alpharogroup/collections/list/ListExtensions.java
@@ -154,6 +154,60 @@ public final class ListExtensions
 	}
 
 	/**
+	 * Gets all possible combinations from the given list
+	 *
+	 * @param <T>
+	 *            the generic type of the elements in the list
+	 * @param numberOf
+	 *            the number of elements to generate the combinations
+	 * @param values
+	 *            the list with the element values
+	 * @return all possible combinations from the given list
+	 */
+	public static <T> List<List<T>> getCombinations(int numberOf, List<T> values)
+	{
+		List<List<T>> combinations = new ArrayList<>();
+		if (numberOf == 0)
+		{
+			combinations.add(new ArrayList<T>());
+			return combinations;
+		}
+		for (int i = 0; i < values.size(); i++)
+		{
+			T element = values.get(i);
+			List<T> rest = getPartialList(values, i + 1);
+			for (List<T> previous : getCombinations(numberOf - 1, rest))
+			{
+				previous.add(element);
+				combinations.add(previous);
+			}
+		}
+		return combinations;
+	}
+
+	/**
+	 * Gets the partial list
+	 *
+	 * @param <T>
+	 *            the generic type of the elements in the list
+	 * @param list
+	 *            the list
+	 * @param i
+	 *            the i
+	 * @return the partial list
+	 */
+	public static <T> List<T> getPartialList(List<T> list, int i)
+	{
+		List<T> partialList = ListFactory.newArrayList();
+		for (int j = i; j < list.size(); j++)
+		{
+			partialList.add(list.get(j));
+		}
+		return partialList;
+	}
+
+
+	/**
 	 * Gets the modified lists. finding from an old list which elements have been removed and which
 	 * have been added.
 	 *
@@ -479,17 +533,18 @@ public final class ListExtensions
 	}
 
 	/**
-	 * Splits the given {@link Collection} to parts to the specified times.
+	 * Splits the given {@link Collection} to parts to the specified size and returns a list with
+	 * the parts.
 	 *
 	 * @param <T>
 	 *            the generic type
 	 * @param collection
 	 *            The collection to split
-	 * @param times
+	 * @param size
 	 *            How to split.
 	 * @return a List with the splited Parts
 	 */
-	public static <T> List<List<T>> splitToParts(final Collection<T> collection, final int times)
+	public static <T> List<List<T>> splitToParts(final Collection<T> collection, final int size)
 	{
 		final List<List<T>> returnList = new ArrayList<>();
 		ArrayList<T> tmp = new ArrayList<>();
@@ -497,7 +552,7 @@ public final class ListExtensions
 		int count = 0;
 		while (it.hasNext())
 		{
-			if (count == times)
+			if (count == size)
 			{
 				returnList.add(tmp);
 				tmp = new ArrayList<>();