+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 ## Change log
 ----------------------
 
+Version 5.1
+-------------
+
+ADDED:
+ 
+- new method created for get all combinations of a given list of values 
+- new method created partition created in java 8 style
+- new method created for remove the first element in a given array
+- new method created for remove the last element in a given array
+- new method created for remove an element over an index in a given array
+
+CHANGED:
+
+- update of parent version to 4.5
+- update of test-objects dependency version to 5.0.1
+
 Version 5
 -------------